+++ fiddles/jquery/fiddle-0042-EniPuzzle/static-app.js
@@ -1,16 +1,6 @@
 (function (app, $, undefined) {
   "use strict";
 
-  app.metadata = {
-    fiddleHeader: 'HTML Drag And Drop API',
-    urls: {
-      github: 'https://github.com/bradyhouse/house/tree/master/fiddles/jquery/fiddle-0043-DragAndDropApi'
-    },
-    consoleTag: 'H O U S E ~ f i d d l e s'
-  };
-
-  app.allowDrop = true;
-
   app.generateGuid = function () {
     return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
       var r = Math.random() * 16 | 0, v = c == 'x' ? r : (r & 0x3 | 0x8);
@@ -21,7 +11,6 @@
   app.onDragStart = function (event) {
     event.dataTransfer.setData('text/plain', event.target.id);
     event.dataTransfer.dropEffect = 'none';
-    app.allowDrop = true;
   };
 
   app.onDragOver = function (event) {
@@ -29,28 +18,11 @@
     event.dataTransfer.dropEffect = 'move'
   };
 
-  app.onAntiDrop = function (event) {
-    app.allowDrop = false;
-    event.preventDefault();
-    event.dataTransfer.clearData();
-    return false;
-  };
-
-  app.onAntiDragOver = function (event) {
-    app.allowDrop = false;
-    event.preventDefault();
-    event.dataTransfer.clearData();
-    event.returnValue = null;
-    return false;
-  };
-
   app.onDrop = function (event) {
-    if (app.allowDrop) {
-      event.preventDefault();
-      let data = event.dataTransfer.getData('text');
-      event.target.appendChild(document.getElementById(data));
-      app.view.createDraggableLink(app.hook);
-    }
+    event.preventDefault();
+    let data = event.dataTransfer.getData('text');
+    event.target.appendChild(document.getElementById(data));
+    app.view.createDraggableLink(app.hook);
   };
 
   app.view = app.view || {
@@ -63,22 +35,12 @@
         el.innerHTML = 'drag to drop - ' + el.id + '<br/>';
         hookEl.appendChild(el);
       },
-      createAntiDropZone: function (hookEl) {
-        let elEmpty = window.document.createElement('div'),
-          el = window.document.createElement('div');
-        el.setAttribute('ondrop', 'app.onAntiDrop(event)');
-        el.setAttribute('ondragover', 'app.onAntiDragOver(event)');
-        el.setAttribute('class', 'anti-drop-zone infinite-slide-left');
-        el.innerHTML = 'anti-drop<br/>';
-        hookEl.appendChild(el);
-        hookEl.appendChild(elEmpty);
-      },
       createDropZone: function (hookEl) {
         let elTitle = window.document.createElement('div'),
           el = window.document.createElement('div');
         el.setAttribute('ondrop', 'app.onDrop(event)');
         el.setAttribute('ondragover', 'app.onDragOver(event)');
-        el.setAttribute('class', 'drop-zone');
+        el.setAttribute('style', 'position: absolute; bottom: 10%; width: 100%; background-color: black; color: white; vertical-align: top; padding: 1%;');
         el.innerHTML = 'drop<br/>';
         hookEl.appendChild(el);
         hookEl.appendChild(elTitle);
@@ -89,18 +51,10 @@
         app.hook = hook;
         app.view.createDropZone(hook);
         app.view.createDraggableLink(hook);
-        app.view.createAntiDropZone(hook);
       }
     };
 
-  app.configConsole = function() {
-    console.log("%c" + app.metadata.consoleTag, 'font-style: italic; font-size: 20px;');
-    console.log("%c" + app.metadata.urls.github, "color: blue; font-style: italic; text-decoration: underline; background-color: #FFFF00;");
-    console.group();
-  }
-
   $(document).ready(function () {
-    app.configConsole();
     app.view.init();
   });